CRUD ok

+++ controllers/stuff.js
@@ -1,30 +1,60 @@
 const Thing = require('../models/Thing');
+const fs = require('fs');
 
 exports.createThing = (req, res, next) => {
-    // Supprime la clé '_id' du corps de la demande
-    delete req.body._id;
-    // Crée un nouvel objet 'Thing' en utilisant les données du corps de la demande
-    const thing = new Thing({
-      ...req.body
-    });
-    // Enregistre l'objet dans la base de données
-    thing.save()
-      // Envoie une réponse avec un code de statut 201 et un message JSON si l'enregistrement est réussi
-      .then(() => res.status(201).json({ message: 'Objet enregistré !'}))
-      // Envoie une réponse avec un code de statut 400 et l'erreur rencontrée si l'enregistrement échoue
-      .catch(error => res.status(400).json({ error }));
+  const thingObject = JSON.parse(req.body.thing);
+  delete thingObject._id;
+  delete thingObject._userId;
+  const thing = new Thing({
+      ...thingObject,
+      userId: req.auth.userId,
+      imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+  });
+
+  thing.save()
+  .then(() => { res.status(201).json({message: 'Objet enregistré !'})})
+  .catch(error => { res.status(400).json( { error })})
 };
 
 exports.modifyThing = (req, res, next) => {
-    Thing.updateOne({ _id: req.params.id}, { ...req.body, _id: req.params.id })
-    .then(() => res.status(200).json({ message: 'Objet modifié !'}))
-    .catch(error => res.status(400).json({ error }));
+  const thingObject = req.file ? {
+    ...JSON.parse(req.body.thing),
+    imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+  } : { ...req.body };
+
+  delete thingObject._userId;
+  Thing.findOne({_id: req.params.id})
+    .then((thing) => {
+      if (thing.userId != req.auth.userId) {
+        res.status(401).json({ message: 'Non-autorisé'});
+      } else {
+        Thing.updateOne({ _id: req.params.id }, { ...thingObject, _id: req.params.id})
+        .then(() => res.status(200).json({message : 'Objet modifié !'}))
+        .catch(error => res.status(401).json({ error }));
+      }
+    })
+    .catch((error) => { 
+      res.status(400).json({ error });
+    });
 };
 
 exports.deleteThing = (req, res, next) => {
-    Thing.deleteOne({ _id: req.params.id })
-      .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
-      .catch(error => res.status(400).json({ error }));
+    Thing.findOne({ _id: req.params.id })
+    .then(thing => {
+      if(thing.userId != req.auth.userId) {
+        res.status(401).json({ message: 'Non-autorisé' });
+      } else {
+        const filename = thing.imageUrl.split('/images/')[1];
+        fs.unlink(`images/${filename}`, () => {
+          Thing.deleteOne({_id: req.params.id})
+          .then(() => { res.status(200).json({ message: 'Objet supprimé !'})})
+          .catch(error => res.status(401).json({ error }));
+        });
+      }
+    })
+    .catch(error => {
+      res.status(500).json({ error });
+    });
 };
 
 exports.getOneThing = (req, res, next) => {
